feat: 편입생 인정 학점을 적용하고 편입생은 채플 1회 수강으로 변경 completed_credit의  total_credit 타입을 int->double

+++ src/main/java/com/plzgraduate/myongjigraduatebe/completedcredit/infrastructure/persistence/entity/CompletedCreditJpaEntity.java
@@ -34,14 +34,14 @@ public class CompletedCreditJpaEntity {
 	@Enumerated(value = EnumType.STRING)
 	private GraduationCategory graduationCategory;
 
-	private int totalCredit;
+	private double totalCredit;
 
 	private double takenCredit;
 
 	@Builder
 	private CompletedCreditJpaEntity(Long id, UserJpaEntity userJpaEntity,
 		GraduationCategory graduationCategory,
-		int totalCredit, double takenCredit) {
+		double totalCredit, double takenCredit) {
 		this.id = id;
 		this.userJpaEntity = userJpaEntity;
 		this.graduationCategory = graduationCategory;
